Generate OpenApi docs with springfox
Add a new component which provides the Docket bean, setting up the OpenAPI
generation and the swagger-ui endpoint.

Add all possible documentation for the currently available API.

Update the Vietnamese endpoint to "vie", since we will standardize on
ISO 639-2 3-letter language codes.

Add a configuration file for the app and set the port to Garfield's birth year.
8080 regularly conflicts with other apps I'm developing on this computer.